requirements.txt
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
multitasking==0.0.9
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
threadpoolctl==2.1.0
urllib3==1.26.6
yfinance==0.1.63
PyYAML==6.0
robin-stocks==2.1.0
python-dotenv==0.19.2
pydantic==1.8.2